+++ src/bindings/nodejs/src/valhalla.cc
@@ -43,13 +43,7 @@ const boost::property_tree::ptree configure(const std::string& config) {
     std::stringstream stream(config);
     rapidjson::read_json(stream, pt);
 
-    auto logging_subtree = pt.get_child_optional("mjolnir.logging");
-    if (logging_subtree) {
-      auto logging_config = valhalla::midgard::ToMap<const boost::property_tree::ptree&,
-                                                     std::unordered_map<std::string, std::string>>(
-          logging_subtree.get());
-      valhalla::midgard::logging::Configure(logging_config);
-    }
+    valhalla::midgard::logging::Configure(pt, "mjolnir.logging");
   } catch (...) { throw std::runtime_error("Failed to load config"); }
 
   return pt;