+++ src/valhalla_service.cc
@@ -195,13 +195,7 @@ int main(int argc, char** argv) {
   }
 
   // configure logging
-  auto logging_subtree = config.get_child_optional("loki.logging");
-  if (logging_subtree) {
-    auto logging_config =
-        valhalla::midgard::ToMap<const boost::property_tree::ptree&,
-                                 std::unordered_map<std::string, std::string>>(logging_subtree.get());
-    valhalla::midgard::logging::Configure(logging_config);
-  }
+  valhalla::midgard::logging::Configure(config, "loki.logging");
 
   // number of workers to use at each stage
   auto worker_concurrency =